refactored MainWindowController
