+++ src/main/java/io/github/mzmine/gui/mainwindow/MainWindowController.java
@@ -1,37 +1,17 @@
-/*
- * Copyright 2006-2020 The MZmine Development Team
- *
- * This file is part of MZmine 3.
- *
- * MZmine 3 is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
- *
- * MZmine 3 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with MZmine 3; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
- * USA
- */
-
 package io.github.mzmine.gui.mainwindow;
 
-import java.text.NumberFormat;
-import java.util.List;
-import java.util.logging.Logger;
-import org.controlsfx.control.StatusBar;
 import com.google.common.collect.Ordering;
 import io.github.mzmine.datamodel.PeakList;
 import io.github.mzmine.datamodel.RawDataFile;
 import io.github.mzmine.gui.MZmineGUI;
+import io.github.mzmine.gui.mainwindow.tabs.processingreport.TextAreaProcessingReportTab;
 import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.modules.visualization.chromatogram.ChromatogramVisualizerModule;
 import io.github.mzmine.modules.visualization.chromatogram.TICVisualizerParameters;
 import io.github.mzmine.modules.visualization.featurelisttable.PeakListTableModule;
 import io.github.mzmine.modules.visualization.fx3d.Fx3DVisualizerModule;
 import io.github.mzmine.modules.visualization.fx3d.Fx3DVisualizerParameters;
+import io.github.mzmine.modules.visualization.rawdataoverview.RawDataOverviewWindowController;
 import io.github.mzmine.modules.visualization.spectra.simplespectra.SpectraVisualizerModule;
 import io.github.mzmine.modules.visualization.spectra.simplespectra.SpectraVisualizerParameters;
 import io.github.mzmine.modules.visualization.twod.TwoDVisualizerModule;
@@ -44,10 +24,14 @@ import io.github.mzmine.taskcontrol.TaskStatus;
 import io.github.mzmine.taskcontrol.impl.WrappedTask;
 import io.github.mzmine.util.ExitCode;
 import io.github.mzmine.util.javafx.FxIconUtil;
+import java.text.NumberFormat;
+import java.util.List;
+import java.util.logging.Logger;
 import javafx.animation.Animation;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.beans.property.ReadOnlyObjectWrapper;
+import javafx.collections.ListChangeListener;
 import javafx.collections.ObservableList;
 import javafx.event.Event;
 import javafx.fxml.FXML;
@@ -57,20 +41,21 @@ import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.ProgressBar;
 import javafx.scene.control.SelectionMode;
+import javafx.scene.control.Tab;
+import javafx.scene.control.TabPane;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.TreeView;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.StackPane;
 import javafx.scene.paint.Color;
 import javafx.util.Duration;
+import org.controlsfx.control.StatusBar;
 
-/**
- * This class controls the main window of the application
- *
- */
 public class MainWindowController {
 
   private final Logger logger = Logger.getLogger(this.getClass().getName());
@@ -90,13 +75,37 @@ public class MainWindowController {
   private Scene mainScene;
 
   @FXML
-  private BorderPane mainWindowPane;
-
-  @FXML
   private ListView<RawDataFile> rawDataTree;
 
   @FXML
   private ListView<PeakList> featureTree;
+
+  @FXML
+  private Tab tvAligned;
+
+  @FXML
+  private TreeView<?> tvAlignedFeatureLists;
+
+  @FXML
+  private AnchorPane tbRawData;
+
+  @FXML
+  private AnchorPane tbFeatureTable;
+
+  @FXML
+  private BorderPane rawDataOverview;
+
+  @FXML
+  private RawDataOverviewWindowController rawDataOverviewController;
+
+  @FXML
+  private TabPane tpMain;
+
+  @FXML
+  private Tab tabRawDataOverview;
+
+  @FXML
+  private Tab tabFeatureListOverview;
 
   @FXML
   private TableView<WrappedTask> tasksView;
@@ -138,6 +147,7 @@ public class MainWindowController {
           return;
         }
         setText(item.getName());
+        setTextFill(item.getColor());
         setGraphic(new ImageView(rawDataFileIcon));
       }
     });
@@ -148,6 +158,15 @@ public class MainWindowController {
         handleShowChromatogram(event);
       }
     });
+
+    rawDataTree.getSelectionModel().getSelectedItems().addListener(
+        new ListChangeListener<>() {
+          @Override
+          public void onChanged(Change<? extends RawDataFile> c) {
+            c.next();
+            rawDataOverviewController.setRawDataFiles((List<RawDataFile>) c.getList());
+          }
+        });
 
     featureTree.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
     // featureTree.setShowRoot(true);
@@ -162,10 +181,11 @@ public class MainWindowController {
           return;
         }
         setText(item.getName());
-        if (item.getNumberOfRawDataFiles() > 1)
+        if (item.getNumberOfRawDataFiles() > 1) {
           setGraphic(new ImageView(featureListAlignedIcon));
-        else
+        } else {
           setGraphic(new ImageView(featureListSingleIcon));
+        }
       }
     });
     // Add mouse clicked event handler
@@ -197,8 +217,9 @@ public class MainWindowController {
       @Override
       public void updateItem(Double value, boolean empty) {
         super.updateItem(value, empty);
-        if (empty)
+        if (empty) {
           return;
+        }
         ProgressBar progressBar = new ProgressBar(value);
         progressBar.setOpacity(0.3);
         progressBar.prefWidthProperty().bind(taskProgressColumn.widthProperty().subtract(20));
@@ -246,6 +267,10 @@ public class MainWindowController {
      * instanceof MZmineTask) { MZmineTask mzmineTask = (MZmineTask) task;
      * mzmineTask.refreshStatus(); } } })); msdkTaskUpdater.play();
      */
+
+    TextAreaProcessingReportTab reportTab = new TextAreaProcessingReportTab("Report: Deisotoping");
+    reportTab.setDemoText();
+    getOverviewTabPane().getTabs().add(reportTab);
   }
 
   public ListView<RawDataFile> getRawDataTree() {
@@ -276,8 +301,9 @@ public class MainWindowController {
     parameters.getParameter(TICVisualizerParameters.DATA_FILES).setValue(
         RawDataFilesSelectionType.SPECIFIC_FILES, selectedFiles.toArray(new RawDataFile[0]));
     ExitCode exitCode = parameters.showSetupDialog(true);
-    if (exitCode == ExitCode.OK)
+    if (exitCode == ExitCode.OK) {
       MZmineCore.runMZmineModule(ChromatogramVisualizerModule.class, parameters);
+    }
   }
 
   public void handleShowMsSpectrum(Event event) {
@@ -288,8 +314,9 @@ public class MainWindowController {
     parameters.getParameter(SpectraVisualizerParameters.dataFiles).setValue(
         RawDataFilesSelectionType.SPECIFIC_FILES, selectedFiles.toArray(new RawDataFile[0]));
     ExitCode exitCode = parameters.showSetupDialog(true);
-    if (exitCode == ExitCode.OK)
+    if (exitCode == ExitCode.OK) {
       MZmineCore.runMZmineModule(SpectraVisualizerModule.class, parameters);
+    }
   }
 
   public void handleShow2DPlot(Event event) {
@@ -300,8 +327,9 @@ public class MainWindowController {
     parameters.getParameter(TwoDVisualizerParameters.dataFiles).setValue(
         RawDataFilesSelectionType.SPECIFIC_FILES, selectedFiles.toArray(new RawDataFile[0]));
     ExitCode exitCode = parameters.showSetupDialog(true);
-    if (exitCode == ExitCode.OK)
+    if (exitCode == ExitCode.OK) {
       MZmineCore.runMZmineModule(TwoDVisualizerModule.class, parameters);
+    }
 
   }
 
@@ -313,30 +341,37 @@ public class MainWindowController {
     parameters.getParameter(Fx3DVisualizerParameters.dataFiles).setValue(
         RawDataFilesSelectionType.SPECIFIC_FILES, selectedFiles.toArray(new RawDataFile[0]));
     ExitCode exitCode = parameters.showSetupDialog(true);
-    if (exitCode == ExitCode.OK)
+    if (exitCode == ExitCode.OK) {
       MZmineCore.runMZmineModule(Fx3DVisualizerModule.class, parameters);
+    }
   }
 
-  public void handleShowMsMsPlot(Event event) {}
+  public void handleShowMsMsPlot(Event event) {
+  }
 
   public void handleSort(Event event) {
-    if (!(event.getSource() instanceof ListView))
+    if (!(event.getSource() instanceof ListView)) {
       return;
+    }
     ListView<?> sourceList = (ListView<?>) event.getSource();
     List<?> files = sourceList.getItems();
     files.sort(Ordering.usingToString());
   }
 
-  public void handleRemoveFileExtension(Event event) {}
+  public void handleRemoveFileExtension(Event event) {
+  }
 
-  public void handleExportFile(Event event) {}
+  public void handleExportFile(Event event) {
+  }
 
-  public void handleRenameFile(Event event) {}
+  public void handleRenameFile(Event event) {
+  }
 
   public void handleRemoveRawData(Event event) {
 
-    if (rawDataTree.getSelectionModel() == null)
+    if (rawDataTree.getSelectionModel() == null) {
       return;
+    }
 
     // Get selected tree items
     ObservableList<RawDataFile> rows = rawDataTree.getSelectionModel().getSelectedItems();
@@ -353,21 +388,26 @@ public class MainWindowController {
 
   public void handleOpenFeatureList(Event event) {
     var selectedFeatureLists = MZmineGUI.getSelectedFeatureLists();
-    for (PeakList fl : selectedFeatureLists)
+    for (PeakList fl : selectedFeatureLists) {
       PeakListTableModule.showNewPeakListVisualizerWindow(fl);
+    }
   }
 
-  public void handleShowFeatureListSummary(Event event) {}
+  public void handleShowFeatureListSummary(Event event) {
+  }
 
-  public void handleShowScatterPlot(Event event) {}
+  public void handleShowScatterPlot(Event event) {
+  }
 
-  public void handleRenameFeatureList(Event event) {}
+  public void handleRenameFeatureList(Event event) {
+  }
 
   @FXML
   public void handleRemoveFeatureList(Event event) {
     PeakList selectedFeatureLists[] = MZmineCore.getDesktop().getSelectedPeakLists();
-    for (PeakList fl : selectedFeatureLists)
+    for (PeakList fl : selectedFeatureLists) {
       MZmineCore.getProjectManager().getCurrentProject().removePeakList(fl);
+    }
   }
 
   @FXML
@@ -410,6 +450,10 @@ public class MainWindowController {
       logger.info("Freeing unused memory");
       System.gc();
     }).start();
+  }
+
+  public TabPane getOverviewTabPane() {
+    return tpMain;
   }
 
 }
